+++ CHANGELOG.md
@@ -52,6 +52,8 @@
 - Remove ProductRepositoryInterface::buildByScope
 - Remove ProductRepositoryInterface::findByExistingFamily
 - Remove ProductRepositoryInterface::findAllByAttributes
+- Move CatalogBundle/Doctrine/ORM/CompletenessJoin and CatalogBundle/Doctrine/ORM/ValueJoin to CatalogBundle/Doctrine/ORM/Join
+- Move CatalogBundle/Doctrine/ORM/CriteriaCondition to CatalogBundle/Doctrine/ORM/Condition
 
 # 1.2.x